Factor out propagation logic
