+++ client-library-otel/src/instrumentation.ts
@@ -12,6 +12,7 @@ import { AsyncLocalStorageContextManager } from "./async-hooks";
 
 import { measure } from "./measure";
 import { patchConsole, patchFetch, patchWaitUntil } from "./patch";
+import { propagateFpxTraceId } from "./propagation";
 import { isRouteInspectorRequest, respondWithRoutes } from "./routes";
 import { getRequestAttributes, getResponseAttributes } from "./utils";
 
@@ -105,28 +106,7 @@ export function instrument(app: Hono, config?: FpxConfigOptions) {
           const promises = patched?.promises ?? [];
           const proxyExecutionCtx = patched?.proxyContext ?? executionContext;
 
-          // HACK - We need to extract the traceparent from the request headers
-          //        but we also need to set a dummy traceparent in the case that
-          //        we are receiving an explicit trace-id to use from FPX.
-          //
-          // Extract fpx trace ID from request headers (if it exists)
-          // and set it as a dummy active context
-          //
-          // TODO - Validate the id here using an otel helper, and warn if it is invalid
-          const traceId = request.headers.get("x-fpx-trace-id");
-          console.log("FPX traceId", traceId);
-
-          let activeContext = context.active();
-          if (traceId) {
-            activeContext = propagation.extract(context.active(), {
-              traceparent: createTraceparentHeader(traceId),
-            });
-          } else {
-            activeContext = propagation.extract(
-              context.active(),
-              request.headers,
-            );
-          }
+          const activeContext = propagateFpxTraceId(request);
 
           const measuredFetch = measure(
             {
@@ -181,6 +161,7 @@ function setupTracerProvider(options: {
   serviceName: string;
   endpoint: string;
 }) {
+  // We need to use async hooks to be able to propagate context
   const asyncHooksContextManager = new AsyncLocalStorageContextManager();
   asyncHooksContextManager.enable();
   context.setGlobalContextManager(asyncHooksContextManager);
@@ -208,30 +189,4 @@ function mergeConfigs(
   return {
     monitor: Object.assign(fallbackConfig.monitor, userConfig?.monitor),
   };
-}
-
-function createTraceparentHeader(traceId: string): string {
-  const version = "00"; // Version of the traceparent header
-  // NOTE - A dummy span id like the following will be rejected by trace propagation API
-  //        Look in the otel codebase for their regex - it filters out anything that's just 0 repeating
-  //
-  // const spanId = '0000000000000000'; // Dummy span ID
-  //
-  // HACK - Generate a random span id so we can spook a proper trace parent
-  //
-  const spanId = generateSpanId();
-  const traceFlags = "01"; // Trace flags (01 means sampled)
-
-  return `${version}-${traceId}-${spanId}-${traceFlags}`;
-}
-
-function generateSpanId(): string {
-  // Generate a random 16-character hex string that is not all zeros
-  let spanId;
-  do {
-    spanId = [...Array(16)]
-      .map(() => Math.floor(Math.random() * 16).toString(16))
-      .join("");
-  } while (/^[0]{16}$/.test(spanId));
-  return spanId;
 }
